Problem19.R
Perhaps this is a little unfair since I'm using R's built-in functionality for handling dates. R not only deals with leap days, but it can also determine the day of the week of a given date using the weekdays() function. This is a nice advantage to using R since it can handle various date types in a sensible manner.