Problem5.R
An example:

Suppose I want to find the least common multiple of the numbers 1:8.

First I find the prime factors for each number. 1 through 3 are prime. 4 is composed of 2^2. 5 is prime. 6 is 3*2. 7 is prime. 8 is 2^3.

Now find the highest power of each prime factor. The highest power of every number in this list is 1 except for 2, whose highest power is 3.

Therefore, we can find the least common multiple by multiplying 1^1*2^3*3^1*5^1*7^1 = 840.