Problem39.R
This deserves some justification.

The available equations to us are (1) a+b+c=P, and (2) a^2+b^2=c^2.

We have three unknowns, so we need three independent equations. We only have two, so we can't analytically solve for solutions given a perimeter P. However, we can solve for a (or b) in terms of b (or a) and P by solving (1) for c and placing it in (2), and then solving for a.

Doing this returns a = (P (2 b-P))/(2 (b-P)) and b!=P. We only care about integer values of a and b, so in the code I check to see if the solution for a, given an integer b, is also an integer. I also don't care about negative solutions, since they don't represent real triangles, and I only accept solutions for which b!=P, which in my case means I reject a=Inf solutions.

Going through the list, it can be seen that starting at a small b and working up returns a large a value which decreases as b increases. Therefore, I can get out of the loop once a is negative or 0, since I know all further solutions will be even more negative. This saves some time.

This solution ends up returning all values for a and b which are solutions to the problem. For example, for p = 120, the solutions returned would be 48, 45, 40, 30, 24, 20. Since all of these numbers are paired with another from the same set, I cut the number of solutions in half to avoid over-counting.

There are ways of optimizing this further - for example, instead of storing a vector that contains all the values, I could only update a variable if the new solution number is larger than the old one, but I like to see what's going on, and looking at the vector m gives insights into how the solutions are populated, how frequently they occur, and what the mode of the solution space is.